% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcium_analysis.R
\name{import_calcium_data}
\alias{import_calcium_data}
\title{Import raw calcium imaging data}
\usage{
import_calcium_data(file)
}
\arguments{
\item{file}{Path to the raw CSV file.}
}
\value{
A list containing:
\describe{
\item{frame}{Frame numbers}
\item{fluorescence_matrix}{Raw fluorescence matrix}
}
}
\description{
Read raw calcium imaging fluorescence traces from a CSV file.
}
